chore: initial & WIP library refactor/reorganization
breaking lib into smaller, better-organized pieces.
