+++ rust/src/mnemonic.rs
@@ -1,0 +1,41 @@
+
+use stack_epic_keychain::mnemonic;
+use rand::thread_rng;
+use std::ffi::CString;
+use std::os::raw::c_char;
+use rand::Rng;
+
+pub fn mnemonic() -> Result<String, mnemonic::Error> {
+    let seed = create_seed(32);
+    match mnemonic::from_entropy(&seed) {
+        Ok(mnemonic_str) => {
+            Ok(mnemonic_str)
+        }, Err(e) => {
+            return  Err(e);
+        }
+    }
+}
+
+pub fn create_seed(seed_length: u64) -> Vec<u8> {
+    let mut seed: Vec<u8> = vec![];
+    let mut rng = thread_rng();
+    for _ in 0..seed_length {
+        seed.push(rng.gen());
+    }
+    seed
+}
+
+pub fn _get_mnemonic() -> Result<*const c_char, mnemonic::Error> {
+    let mut wallet_phrase = "".to_string();
+    match mnemonic() {
+        Ok(phrase) => {
+            wallet_phrase.push_str(&phrase);
+        },Err(e) => {
+            return Err(e);
+        }
+    }
+    let s = CString::new(wallet_phrase).unwrap();
+    let p = s.as_ptr(); // Get a pointer to the underlaying memory for s
+    std::mem::forget(s); // Give up the responsibility of cleaning up/freeing s
+    Ok(p)
+}
